Cleanup
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@856 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ dcpp/Text.h
@@ -137,7 +137,7 @@ namespace Text {
 	tstring tformat(const tstring& src, T t) {
 		tstring ret(src.size() + 64, _T('\0'));
 		int n = _sntprintf(&ret[0], ret.size(), src.c_str(), t);
-		if(n != -1 && n < ret.size()) {
+		if(n != -1 && n < static_cast<int>(ret.size())) {
 			ret.resize(n);
 		}
 		return ret;
@@ -146,7 +146,7 @@ namespace Text {
 	tstring tformat(const tstring& src, T t, T2 t2, T3 t3) {
 		tstring ret(src.size() + 128, _T('\0'));
 		int n = _sntprintf(&ret[0], ret.size(), src.c_str(), t, t2, t3);
-		if(n != -1 && n < ret.size()) {
+		if(n != -1 && n < static_cast<int>(ret.size())) {
 			ret.resize(n);
 		}
 		return ret;
